Set/Get credentials by Portal API Key instead of IP
The portal's IP may differ to the IP seen when starting the test connection. This seems to happen if there's a sabnzbd being proxied to IP B, but tests get sent out by the primary IP A, causing a mismatch and resulting in an empty credentials list being returned.

By matching by the portal's API key we can mitigate this problem. The NNTP server does not know what port the Web UI portal is so it's assuming 8081. If it errors when trying to get the api key it will just revert back to matching by ip.